chore: move domain folder into server
