+++ server/api/namespace.go
@@ -3,7 +3,7 @@ package api
 import (
 	"context"
 
-	"github.com/odpf/stencil/domain"
+	"github.com/odpf/stencil/server/domain"
 	stencilv1beta1 "github.com/odpf/stencil/server/odpf/stencil/v1beta1"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
