+++ server/schema/models.go
@@ -1,7 +1,7 @@
 package schema
 
 import (
-	"github.com/odpf/stencil/domain"
+	"github.com/odpf/stencil/server/domain"
 )
 
 type ParsedSchema interface {
