+++ server/schema/service.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/google/uuid"
-	"github.com/odpf/stencil/domain"
+	"github.com/odpf/stencil/server/domain"
 	"github.com/odpf/stencil/storage"
 )
